cap nhat

+++ font-end-view/src/app/component/checkout/detail-checkout/detail-checkout.component.ts
@@ -15,13 +15,10 @@ import { forkJoin } from 'rxjs';
 })
 export class DetailCheckoutComponent implements OnInit {
   listCart: any = [];
-
   order: any;
-
+  orderCurrent: any;
   statusPayment: any;
-
   email: any;
-
   user: any = {
     id: null,
     code: null,
@@ -37,19 +34,16 @@ export class DetailCheckoutComponent implements OnInit {
     private emailService: EmailService,
     public utilService: UtilService,
     private cartService: CartService,
-    private orderService: OrderService
+    private orderService: OrderService,
   ) {
     const local = JSON.parse(localStorage.getItem('order-bill'));
     this.order = local.order;
     this.listCart = local.listCart;
     this.email = this.order.email;
     this.statusPayment = this.route.snapshot.queryParamMap.get('vnp_TransactionStatus');
-    // sessionStorage.removeItem('order');
-
     this.cartService.updateTotalProducts(0);
 
     const storedUserString = localStorage.getItem('customer');
-
     if (storedUserString) {
       const storedUser = JSON.parse(storedUserString);
       this.user = {
@@ -63,106 +57,12 @@ export class DetailCheckoutComponent implements OnInit {
   }
 
   ngOnInit(): void {
-    // if (this.order.paymentType === 1) {
-    //   if (this.user.id === null) {
-    //     this.orderService.createOrderNotLogin(this.order).subscribe(res => {
-    //       if (res.status === 'OK') {
-    //         this.order = res.data;
-
-    //         const orderDetailPromises = this.listCart.map((item: { productDetailDTO: { id: any; price: any; }; quantity: any; }) => {
-    //           const obj = {
-    //             idOrder: res.data.id,
-    //             idProductDetail: item.productDetailDTO.id,
-    //             quantity: item.quantity,
-    //             // price: item.productDTO?.reducePrice != null || item.productDTO?.percentageReduce != null ? (item.productDTO.price - item.productDTO.reducePrice) : item.productDTO.price,
-    //             price: item.productDetailDTO.price,
-    //             // codeDiscount: this.extractCodeDiscount(item.productDTO),
-    //           };
-    //           return this.orderDetailService.createOrderDetail(obj).toPromise();
-    //         });
-
-    //         forkJoin(orderDetailPromises).subscribe(
-    //           orderDetailsResponses => {
-    //             this.emailService.sendEmailNotLogin(this.order).subscribe(result => {
-    //               localStorage.removeItem('order-bill');
-    //               this.cookieService.delete('cart', '/');
-    //               this.cookieService.delete('checkout', '/');
-    //             });
-
-    //           },
-    //         );
-    //       }
-    //     });
-    //   }
-    //   // else {
-    //   //   this.orderService.createOrder(this.order).subscribe(res => {
-    //   //     if (res.status === 'OK') {
-    //   //       this.order = res.data;
-    //   //       const orderDetailsObservables = this.listCart.map(item => {
-    //   //         const obj = {
-    //   //           idOrder: res.data.id,
-    //   //           idProductDetail: item.productDetailDTO.id,
-    //   //           quantity: item.quantity,
-    //   //           price: item.productDTO?.reducePrice != null || item.productDTO?.percentageReduce != null ? (item.productDTO.price - item.productDTO.reducePrice) : item.productDTO.price,
-    //   //           codeDiscount: this.extractCodeDiscount(item.productDTO),
-    //   //         };
-    //   //         return this.orderDetailService.createOrderDetail(obj);
-    //   //       });
-    //   //       forkJoin(orderDetailsObservables).subscribe(
-    //   //         orderDetailsResponses => {
-    //   //           console.log(orderDetailsResponses);
-    //   //           this.emailService.sendEmail(this.order).subscribe(result => {
-    //   //             console.log('Send email thành công ');
-    //   //           });
-    //   //           localStorage.removeItem('order-bill');
-    //   //           this.cookieService.delete('cart', '/');
-    //   //           this.cookieService.delete('checkout', '/');
-    //   //         },
-    //   //         error => {
-    //   //         }
-    //   //       );
-    //   //     }
-    //   //   });
-    //   // }
-
-    //   // const orderDetailPromises = this.listCart.map(item => {
-    //   //   const obj = {
-    //   //     idOrder: this.order.id,
-    //   //     idProductDetail: item.productDetailDTO.id,
-    //   //     quantity: item.quantity,
-    //   //     price: item.productDTO?.reducePrice != null || item.productDTO?.percentageReduce != null ? (item.productDTO.price - item.productDTO.reducePrice) : item.productDTO.price,
-    //   //     codeDiscount: item.productDTO?.reducePrice != null || item.productDTO?.percentageReduce != null ? item.productDTO.codeDiscount : null,
-    //   //   };
-    //   //   return this.orderDetailService.createOrderDetail(obj).toPromise();
-    //   // });
-
-    //   // Promise.all(orderDetailPromises)
-    //   //   .then(() => {
-    //   //     if (this.email === null || this.email === undefined) {
-    //   //       this.emailService.sendEmail(this.order).subscribe(res => {
-    //   //       });
-    //   //     } else {
-    //   //       const obj = {
-    //   //         ...this.order,
-    //   //         email: this.email
-    //   //       };
-    //   //       this.emailService.sendEmailNotLogin(obj).subscribe(res => {
-    //   //       });
-    //   //     }
-    //   //   })
-    //   //   .catch(error => {
-    //   //     console.error('Error creating order details:', error);
-    //   //   })
-    //   //   .finally(() => {
-    //   //   });
-    // }
-
-    if (this.order.paymentType === 0) {
+    if (this.order.paymentType === 1) {
       if (this.user.id === null) {
         this.orderService.createOrderNotLogin(this.order).subscribe(res => {
           if (res.status === 'OK') {
-            this.order = res.data;
-
+            this.order.code=res.data.code
+            this.orderCurrent = res.data;
             const orderDetailPromises = this.listCart.map((item: { productDetailDTO: { id: any; price: any; }; quantity: any; }) => {
               const obj = {
                 idOrder: res.data.id,
@@ -174,15 +74,13 @@ export class DetailCheckoutComponent implements OnInit {
               };
               return this.orderDetailService.createOrderDetail(obj).toPromise();
             });
-
             forkJoin(orderDetailPromises).subscribe(
               orderDetailsResponses => {
-                this.emailService.sendEmailNotLogin(this.order).subscribe(result => {
+                this.emailService.sendEmailNotLogin(this.orderCurrent).subscribe(result => {
                   localStorage.removeItem('order-bill');
                   this.cookieService.delete('cart', '/');
                   this.cookieService.delete('checkout', '/');
                 });
-
               },
             );
           }
@@ -191,7 +89,8 @@ export class DetailCheckoutComponent implements OnInit {
       else {
         this.orderService.createOrder(this.order).subscribe(res => {
           if (res.status === 'OK') {
-            this.order = res.data;
+            this.order.code=res.data.code
+            this.orderCurrent = res.data;
             const orderDetailPromises = this.listCart.map((item: { productDetailDTO: { id: any; price: any; }; quantity: any; }) => {
               const obj = {
                 idOrder: res.data.id,
@@ -205,7 +104,66 @@ export class DetailCheckoutComponent implements OnInit {
             });
             forkJoin(orderDetailPromises).subscribe(
               orderDetailsResponses => {
-                this.emailService.sendEmail(this.order).subscribe(result => {
+                this.emailService.sendEmail(this.orderCurrent).subscribe(result => {
+                  localStorage.removeItem('order-bill');
+                  this.cookieService.delete('cart', '/');
+                  this.cookieService.delete('checkout', '/');
+                });
+              }
+            );
+          }
+        });
+      }
+    }
+
+    if (this.order.paymentType === 0) {
+      if (this.user.id === null) {
+        this.orderService.createOrderNotLogin(this.order).subscribe(res => {
+          if (res.status === 'OK') {
+            this.order.code=res.data.code
+            this.orderCurrent = res.data;
+            const orderDetailPromises = this.listCart.map((item: { productDetailDTO: { id: any; price: any; }; quantity: any; }) => {
+              const obj = {
+                idOrder: res.data.id,
+                idProductDetail: item.productDetailDTO.id,
+                quantity: item.quantity,
+                // price: item.productDTO?.reducePrice != null || item.productDTO?.percentageReduce != null ? (item.productDTO.price - item.productDTO.reducePrice) : item.productDTO.price,
+                price: item.productDetailDTO.price,
+                // codeDiscount: this.extractCodeDiscount(item.productDTO),
+              };
+              return this.orderDetailService.createOrderDetail(obj).toPromise();
+            });
+            forkJoin(orderDetailPromises).subscribe(
+              orderDetailsResponses => {
+                this.emailService.sendEmailNotLogin(this.orderCurrent).subscribe(result => {
+                  localStorage.removeItem('order-bill');
+                  this.cookieService.delete('cart', '/');
+                  this.cookieService.delete('checkout', '/');
+                });
+              },
+            );
+          }
+        });
+      }
+      else {
+        this.orderService.createOrder(this.order).subscribe(res => {
+          if (res.status === 'OK') {
+            this.order.code=res.data.code
+            this.orderCurrent = res.data;
+            const orderDetailPromises = this.listCart.map((item: { productDetailDTO: { id: any; price: any; }; quantity: any; }) => {
+              const obj = {
+                idOrder: res.data.id,
+                idProductDetail: item.productDetailDTO.id,
+                quantity: item.quantity,
+                // price: item.productDTO?.reducePrice != null || item.productDTO?.percentageReduce != null ? (item.productDTO.price - item.productDTO.reducePrice) : item.productDTO.price,
+                price: item.productDetailDTO.price,
+                // codeDiscount: this.extractCodeDiscount(item.productDTO),
+              };
+              return this.orderDetailService.createOrderDetail(obj).toPromise();
+            });
+            forkJoin(orderDetailPromises).subscribe(
+              orderDetailsResponses => {
+                this.emailService.sendEmail(this.orderCurrent).subscribe(result => {
                   localStorage.removeItem('order-bill');
                   this.cookieService.delete('cart', '/');
                   this.cookieService.delete('checkout', '/');
